refactor: rename d -> distance
